Adding more PATCH endpoints to WizardController

+++ src/main/kotlin/magic/potion/shop/model/Bottle.kt
@@ -27,7 +27,7 @@ data class Bottle(
     @Column(name = "volume")
     @Min(0, message = "If the volume is 0, the bottle is empty")
     @Max(100, message = "If the volume is 100, the bottle is full")
-    var volume: Long = 0
+    var volume: Int = 0
 ) : RepresentationModel<Bottle>() {
 
     constructor() : this(0, Wizard(), null, 0)
